refactor: move session_time_range to a neutral games/formatting.py
session_time_range lived in games/views/session.py, forcing game.py and a test to
import across the view layer (view→view). Move it to a new games/formatting.py
(display formatting, depends only on games.models + common.time), so callers
import from a neutral module and the cross-view import is gone. Drop the now-unused
local_strftime/timeformat from session.py's common.time import.
